:truck: move schema specific methods to own module
